requirements.txt
click>=8.1.7
colorama>=0.4.6
et-xmlfile>=1.1.0
numpy>=1.26.2
openpyxl>=3.2.0b1
pandas>=2.1.3
python-dateutil>=2.8.2
pytz>=2023.3.post1
setuptools>=68.2.2
six>=1.16.0
tzdata>=2023.3
PyYAML>=6.0.1
pathvalidate>=3.2.0